requirements.txt
cassandra_driver==3.27.0
